+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed another odd libcxx issue between miniconda and conda-forge (involving `tuple`)
+- Use fork of `ffnet` with fix for `setup.py` issue (see https://github.com/mrkwjc/ffnet/pull/11)
 
 ### Changed